+++ src/3rdparty/kimageformats/psd.cpp
@@ -3,7 +3,7 @@
 
     SPDX-FileCopyrightText: 2003 Ignacio Castaño <castano@ludicon.com>
     SPDX-FileCopyrightText: 2015 Alex Merry <alex.merry@kde.org>
-    SPDX-FileCopyrightText: 2022-2023 Mirco Miranda <mircomir@outlook.com>
+    SPDX-FileCopyrightText: 2022-2024 Mirco Miranda <mircomir@outlook.com>
 
     SPDX-License-Identifier: LGPL-2.0-or-later
 */
@@ -11,29 +11,24 @@
 /*
  * The early version of this code was based on Thacher Ulrich PSD loading code
  * released into the public domain. See: http://tulrich.com/geekstuff/
- */
-
-/*
+ *
  * Documentation on this file format is available at
  * http://www.adobe.com/devnet-apps/photoshop/fileformatashtml/
- */
-
-/*
+ *
  * Limitations of the current code:
  * - Color spaces other than RGB/Grayscale cannot be read due to lack of QImage
  *   support. Where possible, a conversion to RGB is done:
  *   - CMYK images are converted using an approximated way that ignores the color
- *     information (ICC profile).
+ *     information (ICC profile) with Qt less than 6.8.
  *   - LAB images are converted to sRGB using literature formulas.
- *   - MULICHANNEL images more than 3 channels are converted as CMYK images.
- *   - DUOTONE images are considered as Grayscale images.
- *
- *   NOTE: The best way to convert between different color spaces is to use a
- *   color management engine (e.g. LittleCMS).
+ *   - MULICHANNEL images with 1 channel are treat as Grayscale images.
+ *   - MULICHANNEL images with more than 1 channels are treat as CMYK images.
+ *   - DUOTONE images are treat as Grayscale images.
  */
 
 #include "fastmath_p.h"
 #include "psd_p.h"
+#include "scanlineconverter_p.h"
 #include "util_p.h"
 
 #include <QDataStream>
@@ -58,10 +53,25 @@ typedef quint8 uchar;
  * At the time I'm writing (07/2022), Gwenview and Krita supports linear sRgb but KDE
  * preview creator does not. This is the why, for now, it is disabled.
  */
-//#define PSD_FAST_LAB_CONVERSION
+// #define PSD_FAST_LAB_CONVERSION
+
+/*
+ * Since Qt version 6.8, the 8-bit CMYK format is natively supported.
+ * If you encounter problems with native CMYK support you can continue to force the plugin to convert
+ * to RGB as in previous versions by defining PSD_NATIVE_CMYK_SUPPORT_DISABLED.
+ */
+// #define PSD_NATIVE_CMYK_SUPPORT_DISABLED
 
 namespace // Private.
 {
+
+#if QT_VERSION < QT_VERSION_CHECK(6, 8, 0) || defined(PSD_NATIVE_CMYK_SUPPORT_DISABLED)
+#   define CMYK_FORMAT QImage::Format_Invalid
+#else
+#   define CMYK_FORMAT QImage::Format_CMYK8888
+#endif
+
+#define NATIVE_CMYK (CMYK_FORMAT != QImage::Format_Invalid)
 
 enum Signature : quint32 {
     S_8BIM = 0x3842494D, // '8BIM'
@@ -96,6 +106,10 @@ enum LayerId : quint32 {
 };
 
 struct PSDHeader {
+    PSDHeader() {
+        memset(this, 0, sizeof(PSDHeader));
+    }
+
     uint signature;
     ushort version;
     uchar reserved[6];
@@ -201,8 +215,7 @@ static qint64 readSize(QDataStream &s, bool psb = false)
         quint32 tmp;
         s >> tmp;
         size = tmp;
-    }
-    else {
+    } else {
         s >> size;
     }
     if (s.status() != QDataStream::Ok) {
@@ -418,8 +431,9 @@ PSDLayerAndMaskSection readLayerAndMaskSection(QDataStream &s, bool isPsb, bool 
     if (s.status() == QDataStream::Ok) {
         for (bool ok = true; ok && !lms.atEnd(isPsb);) {
             auto al = readAdditionalLayer(s, &ok);
-            if (ok)
+            if (ok) {
                 lms.additionalLayerInfo.insert(al.id, al);
+            }
         }
     }
 
@@ -455,8 +469,7 @@ PSDColorModeDataSection readColorModeDataSection(QDataStream &s, bool *ok = null
         cms.duotone.data = s.device()->read(size);
         if (cms.duotone.data.size() != size)
             *ok = false;
-    }
-    else {              // read the palette (768 bytes)
+    } else { // read the palette (768 bytes)
         auto&& palette = cms.palette;
         QList<quint8> vect(size);
         for (auto&& v : vect)
@@ -477,7 +490,7 @@ PSDColorModeDataSection readColorModeDataSection(QDataStream &s, bool *ok = null
  */
 static bool setColorSpace(QImage& img, const PSDImageResourceSection& irs)
 {
-    if (!irs.contains(IRI_ICCPROFILE))
+    if (!irs.contains(IRI_ICCPROFILE) || img.isNull())
         return false;
     auto irb = irs.value(IRI_ICCPROFILE);
     auto cs = QColorSpace::fromIccProfile(irb.data);
@@ -505,7 +518,7 @@ static bool setXmpData(QImage& img, const PSDImageResourceSection& irs)
     // NOTE: "XML:com.adobe.xmp" is the meta set by Qt reader when an
     //       XMP packet is found (e.g. when reading a PNG saved by Photoshop).
     //       I'm reusing the same key because a programs could search for it.
-    img.setText(QStringLiteral("XML:com.adobe.xmp"), xmp);
+    img.setText(QStringLiteral(META_KEY_XMP_ADOBE), xmp);
     return true;
 }
 
@@ -606,7 +619,7 @@ static QDataStream &operator>>(QDataStream &s, PSDHeader &header)
 static bool IsValid(const PSDHeader &header)
 {
     if (header.signature != 0x38425053) { // '8BPS'
-        //qDebug() << "PSD header: invalid signature" << header.signature;
+        // qDebug() << "PSD header: invalid signature" << header.signature;
         return false;
     }
     if (header.version != 1 && header.version != 2) {
@@ -669,10 +682,6 @@ static bool IsSupported(const PSDHeader &header)
         header.color_mode != CM_BITMAP) {
         return false;
     }
-    if (header.color_mode == CM_MULTICHANNEL &&
-        header.channel_count < 3) {
-        return false;
-    }
     return true;
 }
 
@@ -704,8 +713,7 @@ qint64 decompress(const char *input, qint64 ilen, char *output, qint64 olen)
                 return -1;
             memcpy(output + j, input + ip, size_t(rr));
             ip += rr;
-        }
-        else if (ip < ilen) {
+        } else if (ip < ilen) {
             rr = qint64(1-n);
             if (available < rr) {
                 --ip;
@@ -733,25 +741,36 @@ static QImage::Format imageFormat(const PSDHeader &header, bool alpha)
     auto format = QImage::Format_Invalid;
     switch(header.color_mode) {
     case CM_RGB:
-        if (header.depth == 32)
+        if (header.depth == 32) {
             format = header.channel_count < 4 || !alpha ? QImage::Format_RGBX32FPx4 : QImage::Format_RGBA32FPx4_Premultiplied;
-        else if (header.depth == 16)
+        } else if (header.depth == 16) {
             format = header.channel_count < 4 || !alpha ? QImage::Format_RGBX64 : QImage::Format_RGBA64_Premultiplied;
-        else
+        } else {
             format = header.channel_count < 4 || !alpha ? QImage::Format_RGB888 : QImage::Format_RGBA8888_Premultiplied;
+        }
         break;
-    case CM_MULTICHANNEL:       // Treat MCH as CMYK (number of channel check is done in IsSupported())
-    case CM_CMYK:               // Photoshop supports CMYK/MCH 8-bits and 16-bits only
-        if (header.depth == 16)
-            format = header.channel_count < 5 || !alpha ? QImage::Format_RGBX64 : QImage::Format_RGBA64;
-        else if (header.depth == 8)
-            format = header.channel_count < 5 || !alpha ? QImage::Format_RGB888 : QImage::Format_RGBA8888;
+    case CM_MULTICHANNEL: // Treat MCH as CMYK or Grayscale
+    case CM_CMYK: // Photoshop supports CMYK/MCH 8-bits and 16-bits only
+        if (NATIVE_CMYK && header.channel_count == 4 && (header.depth == 16 || header.depth == 8)) {
+            format = CMYK_FORMAT;
+        } else if (header.depth == 16) {
+            if (header.channel_count == 1)
+                format = QImage::Format_Grayscale16;
+            else
+                format = header.channel_count < 5 || !alpha ? QImage::Format_RGBX64 : QImage::Format_RGBA64;
+        } else if (header.depth == 8) {
+            if (header.channel_count == 1)
+                format = QImage::Format_Grayscale8;
+            else
+                format = header.channel_count < 5 || !alpha ? QImage::Format_RGB888 : QImage::Format_RGBA8888;
+        }
         break;
-    case CM_LABCOLOR:           // Photoshop supports LAB 8-bits and 16-bits only
-        if (header.depth == 16)
+    case CM_LABCOLOR: // Photoshop supports LAB 8-bits and 16-bits only
+        if (header.depth == 16) {
             format = header.channel_count < 4 || !alpha ? QImage::Format_RGBX64 : QImage::Format_RGBA64;
-        else if (header.depth == 8)
+        } else if (header.depth == 8) {
             format = header.channel_count < 4 || !alpha ? QImage::Format_RGB888 : QImage::Format_RGBA8888;
+        }
         break;
     case CM_GRAYSCALE:
     case CM_DUOTONE:
@@ -845,6 +864,18 @@ inline void planarToChunchy(uchar *target, const char *source, qint32 width, qin
 }
 
 template<class T>
+inline void planarToChunchyCMYK(uchar *target, const char *source, qint32 width, qint32 c, qint32 cn)
+{
+    auto s = reinterpret_cast<const T*>(source);
+    auto t = reinterpret_cast<quint8*>(target);
+    const T d = std::numeric_limits<T>::max() / std::numeric_limits<quint8>::max();
+    for (qint32 x = 0; x < width; ++x) {
+        t[x * cn + c] = quint8((std::numeric_limits<T>::max() - xchg(s[x])) / d);
+    }
+}
+
+
+template<class T>
 inline void planarToChunchyFloatToUInt16(uchar *target, const char *source, qint32 width, qint32 c, qint32 cn)
 {
     auto s = reinterpret_cast<const T*>(source);
@@ -874,16 +905,14 @@ inline void premulConversion(char *stride, qint32 width, qint32 ac, qint32 cn, c
                 auto alpha = *(s + xcn + ac);
                 *(s + xcn + c) = *(s + xcn + c) + alpha - max;
             }
-        }
-        else if (conv == PremulConversion::PS2A || (conv == PremulConversion::PSLab2A && c == 0)) {
+        } else if (conv == PremulConversion::PS2A || (conv == PremulConversion::PSLab2A && c == 0)) {
             for (qint32 x = 0; x < width; ++x) {
                 auto xcn = x * cn;
                 auto alpha = *(s + xcn + ac);
                 if (alpha > 0)
                     *(s + xcn + c) = ((*(s + xcn + c) + alpha - max) * max + alpha / 2) / alpha;
             }
-        }
-        else if (conv == PremulConversion::PSLab2A) {
+        } else if (conv == PremulConversion::PSLab2A) {
             for (qint32 x = 0; x < width; ++x) {
                 auto xcn = x * cn;
                 auto alpha = *(s + xcn + ac);
@@ -904,6 +933,19 @@ inline void monoInvert(uchar *target, const char* source, qint32 bytes)
 }
 
 template<class T>
+inline void rawChannelsCopyToCMYK(uchar *target, qint32 targetChannels, const char *source, qint32 sourceChannels, qint32 width)
+{
+    auto s = reinterpret_cast<const T*>(source);
+    auto t = reinterpret_cast<quint8*>(target);
+    const T d = std::numeric_limits<T>::max() / std::numeric_limits<quint8>::max();
+    for (qint32 c = 0, cs = std::min(targetChannels, sourceChannels); c < cs; ++c) {
+        for (qint32 x = 0; x < width; ++x) {
+            t[x * targetChannels + c] = (std::numeric_limits<T>::max() - s[x * sourceChannels + c]) / d;
+        }
+    }
+}
+
+template<class T>
 inline void rawChannelsCopy(uchar *target, qint32 targetChannels, const char *source, qint32 sourceChannels, qint32 width)
 {
     auto s = reinterpret_cast<const T*>(source);
@@ -916,6 +958,17 @@ inline void rawChannelsCopy(uchar *target, qint32 targetChannels, const char *so
 }
 
 template<class T>
+inline void rawChannelCopy(uchar *target, qint32 targetChannels, qint32 targetChannel, const char *source, qint32 sourceChannels, qint32 sourceChannel, qint32 width)
+{
+    auto s = reinterpret_cast<const T*>(source);
+    auto t = reinterpret_cast<T*>(target);
+    for (qint32 x = 0; x < width; ++x) {
+        t[x * targetChannels + targetChannel] = s[x * sourceChannels + sourceChannel];
+    }
+}
+
+
+template<class T>
 inline void cmykToRgb(uchar *target, qint32 targetChannels, const char *source, qint32 sourceChannels, qint32 width, bool alpha = false)
 {
     auto s = reinterpret_cast<const T*>(source);
@@ -923,22 +976,22 @@ inline void cmykToRgb(uchar *target, qint32 targetChannels, const char *source, 
     auto max = double(std::numeric_limits<T>::max());
     auto invmax = 1.0 / max; // speed improvements by ~10%
 
-    if (sourceChannels < 3) {
-        qDebug() << "cmykToRgb: image is not a valid CMY/CMYK!";
+    if (sourceChannels < 2) {
+        qDebug() << "cmykToRgb: image is not a valid MCH/CMYK!";
         return;
     }
 
     for (qint32 w = 0; w < width; ++w) {
         auto ps = s + sourceChannels * w;
         auto C = 1 - *(ps + 0) * invmax;
-        auto M = 1 - *(ps + 1) * invmax;
-        auto Y = 1 - *(ps + 2) * invmax;
+        auto M = sourceChannels > 1 ? 1 - *(ps + 1) * invmax : 0.0;
+        auto Y = sourceChannels > 2 ? 1 - *(ps + 2) * invmax : 0.0;
         auto K = sourceChannels > 3 ? 1 - *(ps + 3) * invmax : 0.0;
 
         auto pt = t + targetChannels * w;
         *(pt + 0) = T(std::min(max - (C * (1 - K) + K) * max + 0.5, max));
-        *(pt + 1) = T(std::min(max - (M * (1 - K) + K) * max + 0.5, max));
-        *(pt + 2) = T(std::min(max - (Y * (1 - K) + K) * max + 0.5, max));
+        *(pt + 1) = targetChannels > 1 ? T(std::min(max - (M * (1 - K) + K) * max + 0.5, max)) : std::numeric_limits<T>::max();
+        *(pt + 2) = targetChannels > 2 ? T(std::min(max - (Y * (1 - K) + K) * max + 0.5, max)) : std::numeric_limits<T>::max();
         if (targetChannels == 4) {
             if (sourceChannels >= 5 && alpha)
                 *(pt + 3) = *(ps + 4);
@@ -1025,8 +1078,7 @@ bool readChannel(QByteArray& target, QDataStream &stream, quint32 compressedSize
         if (decompress(tmp.data(), tmp.size(), target.data(), target.size()) < 0) {
             return false;
         }
-    }
-    else if (stream.readRawData(target.data(), target.size()) != target.size()) {
+    } else if (stream.readRawData(target.data(), target.size()) != target.size()) {
         return false;
     }
 
@@ -1103,6 +1155,7 @@ static bool LoadPSD(QDataStream &stream, const PSDHeader &header, QImage &img)
     auto imgChannels = imageChannels(img.format());
     auto channel_num = std::min(qint32(header.channel_count), imgChannels);
     auto raw_count = qsizetype(header.width * header.depth + 7) / 8;
+    auto native_cmyk = img.format() == CMYK_FORMAT;
 
     if (header.height > kMaxQVectorSize / header.channel_count / sizeof(quint32)) {
         qWarning() << "LoadPSD() header height/channel_count too big" << header.height << header.channel_count;
@@ -1138,13 +1191,25 @@ static bool LoadPSD(QDataStream &stream, const PSDHeader &header, QImage &img)
 
     // clang-format off
     // checks the need of color conversion (that requires random access to the image)
-    auto randomAccess = (header.color_mode == CM_CMYK) ||
+    auto randomAccess = (header.color_mode == CM_CMYK && !native_cmyk) ||
+                        (header.color_mode == CM_MULTICHANNEL && header.channel_count != 1 && !native_cmyk) ||
                         (header.color_mode == CM_LABCOLOR) ||
-                        (header.color_mode == CM_MULTICHANNEL) ||
                         (header.color_mode != CM_INDEXED && img.hasAlphaChannel());
     // clang-format on
 
     if (randomAccess) {
+        // CMYK with spots (e.g. CMYKA) ICC conversion to RGBA/RGBX
+        QImage tmpCmyk;
+        ScanLineConverter iccConv(img.format());
+#if QT_VERSION >= QT_VERSION_CHECK(6, 8, 0) && !defined(PSD_NATIVE_CMYK_SUPPORT_DISABLED)
+        if (header.color_mode == CM_CMYK && img.format() != QImage::Format_CMYK8888) {
+            auto tmpi = QImage(header.width, 1, QImage::Format_CMYK8888);
+            if (setColorSpace(tmpi, irs))
+                tmpCmyk = tmpi;
+            iccConv.setTargetColorSpace(QColorSpace(QColorSpace::SRgb));
+        }
+#endif
+
         // In order to make a colorspace transformation, we need all channels of a scanline
         QByteArray psdScanline;
         psdScanline.resize(qsizetype(header.width * header.depth * header.channel_count + 7) / 8);
@@ -1164,11 +1229,9 @@ static bool LoadPSD(QDataStream &stream, const PSDHeader &header, QImage &img)
                 auto scanLine = reinterpret_cast<unsigned char*>(psdScanline.data());
                 if (header.depth == 8) {
                     planarToChunchy<quint8>(scanLine, rawStride.data(), header.width, c, header.channel_count);
-                }
-                else if (header.depth == 16) {
+                } else if (header.depth == 16) {
                     planarToChunchy<quint16>(scanLine, rawStride.data(), header.width, c, header.channel_count);
-                }
-                else if (header.depth == 32) {
+                } else if (header.depth == 32) {
                     planarToChunchy<float>(scanLine, rawStride.data(), header.width, c, header.channel_count);
                 }
             }
@@ -1200,10 +1263,24 @@ static bool LoadPSD(QDataStream &stream, const PSDHeader &header, QImage &img)
 
             // Conversion to RGB
             if (header.color_mode == CM_CMYK || header.color_mode == CM_MULTICHANNEL) {
-                if (header.depth == 8)
-                    cmykToRgb<quint8>(img.scanLine(y), imgChannels, psdScanline.data(), header.channel_count, header.width, alpha);
-                else if (header.depth == 16)
-                    cmykToRgb<quint16>(img.scanLine(y), imgChannels, psdScanline.data(), header.channel_count, header.width, alpha);
+                if (tmpCmyk.isNull()) {
+                    if (header.depth == 8)
+                        cmykToRgb<quint8>(img.scanLine(y), imgChannels, psdScanline.data(), header.channel_count, header.width, alpha);
+                    else if (header.depth == 16)
+                        cmykToRgb<quint16>(img.scanLine(y), imgChannels, psdScanline.data(), header.channel_count, header.width, alpha);
+                } else if (header.depth == 8) {
+                    rawChannelsCopyToCMYK<quint8>(tmpCmyk.bits(), 4, psdScanline.data(), header.channel_count, header.width);
+                    if (auto rgbPtr = iccConv.convertedScanLine(tmpCmyk, 0))
+                        std::memcpy(img.scanLine(y), rgbPtr, img.bytesPerLine());
+                    if (imgChannels == 4 && header.channel_count >= 5)
+                        rawChannelCopy<quint8>(img.scanLine(y), imgChannels, 3, psdScanline.data(), header.channel_count, 4, header.width);
+                } else if (header.depth == 16) {
+                    rawChannelsCopyToCMYK<quint16>(tmpCmyk.bits(), 4, psdScanline.data(), header.channel_count, header.width);
+                    if (auto rgbPtr = iccConv.convertedScanLine(tmpCmyk, 0))
+                        std::memcpy(img.scanLine(y), rgbPtr, img.bytesPerLine());
+                    if (imgChannels == 4 && header.channel_count >= 5)
+                        rawChannelCopy<quint16>(img.scanLine(y), imgChannels, 3, psdScanline.data(), header.channel_count, 4, header.width);
+                }
             }
             if (header.color_mode == CM_LABCOLOR) {
                 if (header.depth == 8)
@@ -1220,8 +1297,7 @@ static bool LoadPSD(QDataStream &stream, const PSDHeader &header, QImage &img)
                     rawChannelsCopy<float>(img.scanLine(y), imgChannels, psdScanline.data(), header.channel_count, header.width);
             }
         }
-    }
-    else {
+    } else {
         // Linear read (no position jumps): optimized code usable only for the colorspaces supported by QImage
         for (qint32 c = 0; c < channel_num; ++c) {
             for (qint32 y = 0, h = header.height; y < h; ++y) {
@@ -1232,25 +1308,31 @@ static bool LoadPSD(QDataStream &stream, const PSDHeader &header, QImage &img)
                 }
 
                 auto scanLine = img.scanLine(y);
-                if (header.depth == 1) { // Bitmap
+                if (header.depth == 1) {
+                    // Bitmap
                     monoInvert(scanLine, rawStride.data(), std::min(rawStride.size(), img.bytesPerLine()));
-                }
-                else if (header.depth == 8) { // 8-bits images: Indexed, Grayscale, RGB/RGBA
-                    planarToChunchy<quint8>(scanLine, rawStride.data(), header.width, c, imgChannels);
-                }
-                else if (header.depth == 16) { // 16-bits integer images: Grayscale, RGB/RGBA
-                    planarToChunchy<quint16>(scanLine, rawStride.data(), header.width, c, imgChannels);
-                }
-                else if (header.depth == 32 && header.color_mode == CM_RGB) { // 32-bits float images: RGB/RGBA
+                } else if (header.depth == 8) {
+                    // 8-bits images: Indexed, Grayscale, RGB/RGBA, CMYK, MCH1, MCH4
+                    if (native_cmyk)
+                        planarToChunchyCMYK<quint8>(scanLine, rawStride.data(), header.width, c, imgChannels);
+                    else
+                        planarToChunchy<quint8>(scanLine, rawStride.data(), header.width, c, imgChannels);
+                } else if (header.depth == 16) {
+                    // 16-bits integer images: Grayscale, RGB/RGBA, CMYK, MCH1, MCH4
+                    if (native_cmyk)
+                        planarToChunchyCMYK<quint16>(scanLine, rawStride.data(), header.width, c, imgChannels);
+                    else
+                        planarToChunchy<quint16>(scanLine, rawStride.data(), header.width, c, imgChannels);
+                } else if (header.depth == 32 && header.color_mode == CM_RGB) {
+                    // 32-bits float images: RGB/RGBA
                     planarToChunchy<float>(scanLine, rawStride.data(), header.width, c, imgChannels);
-                }
-                else if (header.depth == 32 && header.color_mode == CM_GRAYSCALE) { // 32-bits float images: Grayscale (coverted to equivalent integer 16-bits)
+                } else if (header.depth == 32 && header.color_mode == CM_GRAYSCALE) {
+                    // 32-bits float images: Grayscale (coverted to equivalent integer 16-bits)
                     planarToChunchyFloatToUInt16<float>(scanLine, rawStride.data(), header.width, c, imgChannels);
                 }
             }
         }
     }
-
 
     // Resolution info
     if (!setResolution(img, irs)) {
@@ -1265,8 +1347,7 @@ static bool LoadPSD(QDataStream &stream, const PSDHeader &header, QImage &img)
 #else
         img.setColorSpace(QColorSpace(QColorSpace::SRgb));
 #endif
-    }
-    else if (!setColorSpace(img, irs)) {
+    } else if (!setColorSpace(img, irs)) {
         // qDebug() << "No colorspace info set!";
     }
 
@@ -1287,7 +1368,17 @@ static bool LoadPSD(QDataStream &stream, const PSDHeader &header, QImage &img)
 
 } // Private
 
+class PSDHandlerPrivate
+{
+public:
+    PSDHandlerPrivate() {}
+    ~PSDHandlerPrivate() {}
+    PSDHeader m_header;
+};
+
 PSDHandler::PSDHandler()
+    : QImageIOHandler()
+    , d(new PSDHandlerPrivate)
 {
 }
 
@@ -1305,24 +1396,24 @@ bool PSDHandler::read(QImage *image)
     QDataStream s(device());
     s.setByteOrder(QDataStream::BigEndian);
 
-    PSDHeader header;
+    auto&& header = d->m_header;
     s >> header;
 
     // Check image file format.
     if (s.atEnd() || !IsValid(header)) {
-        //         qDebug() << "This PSD file is not valid.";
+        // qDebug() << "This PSD file is not valid.";
         return false;
     }
 
     // Check if it's a supported format.
     if (!IsSupported(header)) {
-        //         qDebug() << "This PSD file is not supported.";
+        // qDebug() << "This PSD file is not supported.";
         return false;
     }
 
     QImage img;
     if (!LoadPSD(s, header, img)) {
-        //         qDebug() << "Error loading PSD file.";
+        // qDebug() << "Error loading PSD file.";
         return false;
     }
 
@@ -1342,18 +1433,15 @@ QVariant PSDHandler::option(ImageOption option) const
     QVariant v;
 
     if (option == QImageIOHandler::Size) {
-        if (auto d = device()) {
-            // transactions works on both random and sequential devices
-            d->startTransaction();
-            auto ba = d->read(sizeof(PSDHeader));
-            d->rollbackTransaction();
-
+        auto&& header = d->m_header;
+        if (IsValid(header)) {
+            v = QVariant::fromValue(QSize(header.width, header.height));
+        } else if (auto dev = device()) {
+            auto ba = dev->peek(sizeof(PSDHeader));
             QDataStream s(ba);
             s.setByteOrder(QDataStream::BigEndian);
 
-            PSDHeader header;
             s >> header;
-
             if (s.status() == QDataStream::Ok && IsValid(header))
                 v = QVariant::fromValue(QSize(header.width, header.height));
         }
@@ -1369,22 +1457,23 @@ bool PSDHandler::canRead(QIODevice *device)
         return false;
     }
 
-    device->startTransaction();
-
-    QDataStream s(device);
+    auto ba = device->peek(sizeof(PSDHeader));
+    QDataStream s(ba);
     s.setByteOrder(QDataStream::BigEndian);
 
     PSDHeader header;
     s >> header;
-
-    device->rollbackTransaction();
 
     if (s.status() != QDataStream::Ok) {
         return false;
     }
 
     if (device->isSequential()) {
-        if (header.color_mode == CM_CMYK || header.color_mode == CM_LABCOLOR || header.color_mode == CM_MULTICHANNEL) {
+        if (header.color_mode == CM_CMYK || header.color_mode == CM_MULTICHANNEL) {
+            if (header.channel_count != 4 || !NATIVE_CMYK)
+                return false;
+        }
+        if (header.color_mode == CM_LABCOLOR) {
             return false;
         }
         if (header.color_mode == CM_RGB && header.channel_count > 3) {