+++ src/3rdparty/kimageformats/util_p.h
@@ -13,6 +13,33 @@
 #include <QImage>
 #include <QImageIOHandler>
 
+// Image metadata keys to use in plugins (so they are consistent)
+#define META_KEY_ALTITUDE "Altitude"
+#define META_KEY_AUTHOR "Author"
+#define META_KEY_COMMENT "Comment"
+#define META_KEY_COPYRIGHT "Copyright"
+#define META_KEY_CREATIONDATE "CreationDate"
+#define META_KEY_DESCRIPTION "Description"
+#define META_KEY_DOCUMENTNAME "DocumentName"
+#define META_KEY_HOSTCOMPUTER "HostComputer"
+#define META_KEY_LATITUDE "Latitude"
+#define META_KEY_LONGITUDE "Longitude"
+#define META_KEY_OWNER "Owner"
+#define META_KEY_SOFTWARE "Software"
+#define META_KEY_TITLE "Title"
+#define META_KEY_XML_GIMP "XML:org.gimp.xml"
+#define META_KEY_XMP_ADOBE "XML:com.adobe.xmp"
+
+// Camera info metadata keys
+#define META_KEY_MANUFACTURER "Manufacturer"
+#define META_KEY_MODEL "Model"
+#define META_KEY_SERIALNUMBER "SerialNumber"
+
+// Lens info metadata keys
+#define META_KEY_LENS_MANUFACTURER "LensManufacturer"
+#define META_KEY_LENS_MODEL "LensModel"
+#define META_KEY_LENS_SERIALNUMBER "LensSerialNumber"
+
 // QList uses some extra space for stuff, hence the 32 here suggested by Thiago Macieira
 static constexpr int kMaxQVectorSize = std::numeric_limits<int>::max() - 32;