+++ src/3rdparty/kimageformats/xcf.cpp
@@ -963,7 +963,7 @@ bool XCFImageFormat::loadImageProperties(QDataStream &xcf_io, XCFImage &xcf_imag
 #if QT_VERSION < QT_VERSION_CHECK(6, 7, 0)
                 quint32 size;
 #else
-                qsizetype size;
+                qint64 size;
 #endif
 
                 property.readBytes(tag, size);
@@ -1091,7 +1091,9 @@ bool XCFImageFormat::loadProperty(QDataStream &xcf_io, PropType &type, QByteArra
         size = 0;
     } else {
         xcf_io >> size;
-        if (size > 256000) {
+        if (size > 256000 * 4) {
+            // NOTE: I didn't find any reference to maximum property dimensions in the specs, so I assume it's just a sanity check.
+            qCDebug(XCFPLUGIN) << "XCF: loadProperty skips" << type << "due to size being too large";
             return false;
         }
         data = new char[size];
@@ -1303,21 +1305,22 @@ bool XCFImageFormat::loadLayerProperties(QDataStream &xcf_io, Layer &layer)
         case PROP_COMPOSITE_SPACE:
             property >> layer.compositeSpace;
             if (layer.compositeSpace < 0) {
-                layer.compositeSpace = GimpColorSpace(-layer.compositeSpace);
+                layer.compositeSpace = GimpColorSpace(layer.compositeSpace == std::numeric_limits<qint32>::lowest() ? 0 : -layer.compositeSpace);
             }
             break;
 
         case PROP_COMPOSITE_MODE:
             property >> layer.compositeMode;
             if (layer.compositeMode < 0) {
-                layer.compositeMode = XCFImageFormat::GimpCompositeMode(-layer.compositeMode);
+                layer.compositeMode =
+                    XCFImageFormat::GimpCompositeMode(layer.compositeMode == std::numeric_limits<qint32>::lowest() ? 0 : -layer.compositeMode);
             }
             break;
 
         case PROP_BLEND_SPACE:
             property >> layer.blendSpace;
-            if (layer.blendSpace) {
-                layer.blendSpace = GimpColorSpace(-layer.blendSpace);
+            if (layer.blendSpace < 0) {
+                layer.blendSpace = GimpColorSpace(layer.blendSpace == std::numeric_limits<qint32>::lowest() ? 0 : -layer.blendSpace);
             }
             break;
 
@@ -1594,7 +1597,7 @@ void XCFImageFormat::setImageParasites(const XCFImage &xcf_image, QImage &image)
         //    comments may also be present in the "gimp-metadata" parasite.
         if (key == QStringLiteral("gimp-comment")) {
             value.replace('\0', QByteArray());
-            image.setText(QStringLiteral("Comment"), QString::fromUtf8(value));
+            image.setText(QStringLiteral(META_KEY_COMMENT), QString::fromUtf8(value));
             continue;
         }
 
@@ -1605,7 +1608,7 @@ void XCFImageFormat::setImageParasites(const XCFImage &xcf_image, QImage &image)
             // NOTE: I arbitrary defined the metadata "XML:org.gimp.xml" because it seems
             //       a GIMP proprietary XML format (no xmlns defined)
             value.replace('\0', QByteArray());
-            image.setText(QStringLiteral("XML:org.gimp.xml"), QString::fromUtf8(value));
+            image.setText(QStringLiteral(META_KEY_XML_GIMP), QString::fromUtf8(value));
             continue;
         }
 
@@ -1621,7 +1624,7 @@ void XCFImageFormat::setImageParasites(const XCFImage &xcf_image, QImage &image)
             //       XMP packet is found (e.g. when reading a PNG saved by Photoshop).
             //       I reused the same key because some programs could search for it.
             value.replace('\0', QByteArray());
-            image.setText(QStringLiteral("XML:com.adobe.xmp"), QString::fromUtf8(value));
+            image.setText(QStringLiteral(META_KEY_XMP_ADOBE), QString::fromUtf8(value));
             continue;
         }
 #endif
@@ -1670,13 +1673,17 @@ bool XCFImageFormat::assignImageBytes(Layer &layer, uint i, uint j, const GimpPr
         for (int y = 0; y < height; y++) {
             uchar *dataPtr = bits + y * bytesPerLine;
             uchar *alphaPtr = nullptr;
-            if (!layer.alpha_tiles.isEmpty())
-                alphaPtr = layer.alpha_tiles[j][i].scanLine(y);
+            if (layer.alpha_tiles.size() > j && layer.alpha_tiles.at(j).size() > i) {
+                QImage &alphaTile = layer.alpha_tiles[j][i];
+                if (alphaTile.width() >= width && alphaTile.height() > y) {
+                    alphaPtr = alphaTile.scanLine(y);
+                }
+            }
             if (bpc == 4) {
 #ifdef USE_FLOAT_IMAGES
                 if (precision < GimpPrecision::GIMP_PRECISION_HALF_LINEAR) {
                     for (int x = 0; x < width; x++) {
-                        auto src = (const quint16 *)tile;
+                        auto src = reinterpret_cast<const quint16 *>(tile);
                         *dataPtr++ = qFromBigEndian<quint16>(src[0]) / 257;
                         if (alphaPtr) {
                             *alphaPtr++ = qFromBigEndian<quint16>(src[1]) / 257;
@@ -1687,7 +1694,7 @@ bool XCFImageFormat::assignImageBytes(Layer &layer, uint i, uint j, const GimpPr
                     }
                 } else {
                     for (int x = 0; x < width; x++) {
-                        auto src = (const float *)tile;
+                        auto src = reinterpret_cast<const float *>(tile);
                         *dataPtr++ = qFromBigEndian<float>(src[0]) * 255;
                         if (alphaPtr) {
                             *alphaPtr++ = qFromBigEndian<float>(src[1]) * 255;
@@ -1713,7 +1720,7 @@ bool XCFImageFormat::assignImageBytes(Layer &layer, uint i, uint j, const GimpPr
 #ifdef USE_FLOAT_IMAGES
                 if (precision < GimpPrecision::GIMP_PRECISION_HALF_LINEAR) {
                     for (int x = 0; x < width; x++) {
-                        auto src = (const quint16 *)tile;
+                        auto src = reinterpret_cast<const quint16 *>(tile);
                         *dataPtr++ = qFromBigEndian<quint16>(src[0]) / 257;
                         if (alphaPtr)
                             *alphaPtr++ = qFromBigEndian<quint16>(src[1]) / 257;
@@ -1721,7 +1728,7 @@ bool XCFImageFormat::assignImageBytes(Layer &layer, uint i, uint j, const GimpPr
                     }
                 } else {
                     for (int x = 0; x < width; x++) {
-                        auto src = (const qfloat16 *)tile;
+                        auto src = reinterpret_cast<const qfloat16 *>(tile);
                         *dataPtr++ = qFromBigEndian<qfloat16>(src[0]) * 255;
                         if (alphaPtr)
                             *alphaPtr++ = qFromBigEndian<qfloat16>(src[1]) * 255;
@@ -1730,7 +1737,7 @@ bool XCFImageFormat::assignImageBytes(Layer &layer, uint i, uint j, const GimpPr
                 }
 #else
                 for (int x = 0; x < width; x++) {
-                    auto src = (const quint16 *)tile;
+                    auto src = reinterpret_cast<const quint16 *>(tile);
                     *dataPtr++ = qFromBigEndian<quint16>(src[0]) / 257;
                     if (alphaPtr)
                         *alphaPtr++ = qFromBigEndian<quint16>(src[1]) / 257;
@@ -1770,9 +1777,9 @@ bool XCFImageFormat::assignImageBytes(Layer &layer, uint i, uint j, const GimpPr
         break;
     case QImage::Format_RGBX64:
         for (int y = 0; y < height; y++) {
-            quint16 *dataPtr = (quint16 *)image.scanLine(y);
+            quint16 *dataPtr = reinterpret_cast<quint16 *>(image.scanLine(y));
             const size_t bpl = width * sizeof(QRgba64);
-            const quint16 *src = (const quint16 *)(tile + y * bpl);
+            const quint16 *src = reinterpret_cast<const quint16 *>(tile + y * bpl);
             for (int x = 0; x < width * 4; x += 4) {
                 dataPtr[x + 0] = qFromBigEndian(src[x + 0]);
                 dataPtr[x + 1] = qFromBigEndian(src[x + 1]);
@@ -1784,8 +1791,8 @@ bool XCFImageFormat::assignImageBytes(Layer &layer, uint i, uint j, const GimpPr
 #ifdef USE_FLOAT_IMAGES
     case QImage::Format_RGBX16FPx4:
         for (int y = 0; y < height; y++) {
-            qfloat16 *dataPtr = (qfloat16 *)image.scanLine(y);
-            const qfloat16 *src = (const qfloat16 *)(tile + y * width * sizeof(QRgbaFloat16));
+            qfloat16 *dataPtr = reinterpret_cast<qfloat16 *>(image.scanLine(y));
+            const qfloat16 *src = reinterpret_cast<const qfloat16 *>(tile + y * width * sizeof(QRgbaFloat16));
             for (int x = 0; x < width * 4; x += 4) {
                 dataPtr[x + 0] = qFromBigEndian(src[x + 0]);
                 dataPtr[x + 1] = qFromBigEndian(src[x + 1]);
@@ -1812,8 +1819,8 @@ bool XCFImageFormat::assignImageBytes(Layer &layer, uint i, uint j, const GimpPr
         break;
     case QImage::Format_RGBX32FPx4:
         for (int y = 0; y < height; y++) {
-            float *dataPtr = (float *)image.scanLine(y);
-            const float *src = (const float *)(tile + y * width * sizeof(QRgbaFloat32));
+            float *dataPtr = reinterpret_cast<float *>(image.scanLine(y));
+            const float *src = reinterpret_cast<const float *>(tile + y * width * sizeof(QRgbaFloat32));
             for (int x = 0; x < width * 4; x += 4) {
                 dataPtr[x + 0] = qFromBigEndian(src[x + 0]);
                 dataPtr[x + 1] = qFromBigEndian(src[x + 1]);
@@ -1953,7 +1960,7 @@ static bool convertFloatTo16Bit(uchar *output, quint64 outputSize, uchar *input)
 {
     SourceFormat *source = (SourceFormat *)(input);
     for (quint64 offset = 0; offset < outputSize; offset++) {
-        ((uint16_t *)output)[offset] = qToBigEndian(quint16(qBound(0., qFromBigEndian<SourceFormat>(source[offset]) * 65535. + 0.5, 65535.)));
+        (reinterpret_cast<uint16_t *>(output))[offset] = qToBigEndian(quint16(qBound(0., qFromBigEndian<SourceFormat>(source[offset]) * 65535. + 0.5, 65535.)));
     }
     return true;
 }
@@ -1968,6 +1975,12 @@ static bool convertFloatTo16Bit(uchar *output, quint64 outputSize, uchar *input)
  */
 bool XCFImageFormat::loadLevel(QDataStream &xcf_io, Layer &layer, qint32 bpp, const GimpPrecision precision)
 {
+    auto bpc = bytesPerChannel(precision);
+    if ((bpc == 0) || (bpp % bpc)) {
+        qCDebug(XCFPLUGIN) << "XCF: the stream seems corrupted";
+        return false;
+    }
+
     qint32 width;
     qint32 height;
 
@@ -2102,9 +2115,9 @@ bool XCFImageFormat::loadLevel(QDataStream &xcf_io, Layer &layer, qint32 bpp, co
                 case GIMP_PRECISION_U32_LINEAR:
                 case GIMP_PRECISION_U32_NON_LINEAR:
                 case GIMP_PRECISION_U32_PERCEPTUAL: {
-                    quint32 *source = (quint32 *)(buffer.data());
+                    quint32 *source = reinterpret_cast<quint32 *>(buffer.data());
                     for (quint64 offset = 0, len = buffer.size() / sizeof(quint32); offset < len; ++offset) {
-                        ((quint16 *)layer.tile)[offset] = qToBigEndian<quint16>(qFromBigEndian(source[offset]) / 65537);
+                        (reinterpret_cast<quint16 *>(layer.tile))[offset] = qToBigEndian<quint16>(qFromBigEndian(source[offset]) / 65537);
                     }
                     break;
                 }
@@ -2128,9 +2141,9 @@ bool XCFImageFormat::loadLevel(QDataStream &xcf_io, Layer &layer, qint32 bpp, co
                 case GIMP_PRECISION_DOUBLE_LINEAR:
                 case GIMP_PRECISION_DOUBLE_NON_LINEAR:
                 case GIMP_PRECISION_DOUBLE_PERCEPTUAL: {
-                    double *source = (double *)(buffer.data());
+                    double *source = reinterpret_cast<double *>(buffer.data());
                     for (quint64 offset = 0, len = buffer.size() / sizeof(double); offset < len; ++offset) {
-                        ((float *)layer.tile)[offset] = qToBigEndian<float>(float(qFromBigEndian(source[offset])));
+                        (reinterpret_cast<float *>(layer.tile))[offset] = qToBigEndian<float>(float(qFromBigEndian(source[offset])));
                     }
                     break;
                 }
@@ -2475,12 +2488,12 @@ bool XCFImageFormat::assignMaskBytes(Layer &layer, uint i, uint j, const GimpPre
         if (bpc == 4) {
             if (precision < GimpPrecision::GIMP_PRECISION_HALF_LINEAR) {
                 for (int x = 0; x < width; x++) {
-                    *dataPtr++ = qFromBigEndian<quint16>(*(const quint16 *)tile) / 257;
+                    *dataPtr++ = qFromBigEndian<quint16>(*reinterpret_cast<const quint16 *>(tile)) / 257;
                     tile += sizeof(quint16); // was converted to 16 bits in loadLevel()
                 }
             } else {
                 for (int x = 0; x < width; x++) {
-                    *dataPtr++ = qFromBigEndian<float>(*(const float *)tile) * 255;
+                    *dataPtr++ = qFromBigEndian<float>(*reinterpret_cast<const float *>(tile)) * 255;
                     tile += sizeof(QRgb); // yeah! see loadTileRLE()
                 }
             }
@@ -2488,12 +2501,12 @@ bool XCFImageFormat::assignMaskBytes(Layer &layer, uint i, uint j, const GimpPre
             // when not converted, the step of a
             if (precision < GimpPrecision::GIMP_PRECISION_HALF_LINEAR) {
                 for (int x = 0; x < width; x++) {
-                    *dataPtr++ = qFromBigEndian<quint16>(*(const quint16 *)tile) / 257;
+                    *dataPtr++ = qFromBigEndian<quint16>(*reinterpret_cast<const quint16 *>(tile)) / 257;
                     tile += sizeof(QRgb); // yeah! see loadTileRLE()
                 }
             } else {
                 for (int x = 0; x < width; x++) {
-                    *dataPtr++ = qFromBigEndian<qfloat16>(*(const qfloat16 *)tile) * 255;
+                    *dataPtr++ = qFromBigEndian<qfloat16>(*reinterpret_cast<const qfloat16 *>(tile)) * 255;
                     tile += sizeof(QRgb); // yeah! see loadTileRLE()
                 }
             }
@@ -2501,12 +2514,12 @@ bool XCFImageFormat::assignMaskBytes(Layer &layer, uint i, uint j, const GimpPre
 #else
         if (bpc == 2) {
             for (int x = 0; x < width; x++) {
-                *dataPtr++ = qFromBigEndian<quint16>(*(const quint16 *)tile) / 257;
+                *dataPtr++ = qFromBigEndian<quint16>(*reinterpret_cast<const quint16 *>(tile)) / 257;
                 tile += sizeof(QRgb); // yeah! see loadTileRLE() / loadLevel()
             }
         } else if (bpc == 4) {
             for (int x = 0; x < width; x++) {
-                *dataPtr++ = qFromBigEndian<quint16>(*(const quint16 *)tile) / 257;
+                *dataPtr++ = qFromBigEndian<quint16>(*reinterpret_cast<const quint16 *>(tile)) / 257;
                 tile += sizeof(quint16); // was converted to 16 bits in loadLevel()
             }
         }
@@ -2753,10 +2766,10 @@ void XCFImageFormat::copyLayerToImage(XCFImage &xcf_image)
     // For each tile...
 
     for (uint j = 0; j < layer.nrows; j++) {
-        uint y = j * TILE_HEIGHT;
+        qint32 y = qint32(j * TILE_HEIGHT);
 
         for (uint i = 0; i < layer.ncols; i++) {
-            uint x = i * TILE_WIDTH;
+            qint32 x = qint32(i * TILE_WIDTH);
 
             // This seems the best place to apply the dissolve because it
             // depends on the global position of each tile's
@@ -3043,7 +3056,7 @@ void XCFImageFormat::mergeLayerIntoImage(XCFImage &xcf_image)
         merge = mergeRGBToRGB;
         break;
     case GRAY_GIMAGE:
-        if (layer.opacity == OPAQUE_OPACITY) {
+        if (layer.opacity == OPAQUE_OPACITY && xcf_image.image.depth() <= 8) {
             merge = mergeGrayToGray;
         } else {
             merge = mergeGrayToRGB;
@@ -3179,10 +3192,10 @@ void XCFImageFormat::mergeLayerIntoImage(XCFImage &xcf_image)
             qCDebug(XCFPLUGIN) << "Using QPainter for mode" << layer.mode;
 
             for (uint j = 0; j < layer.nrows; j++) {
-                uint y = j * TILE_HEIGHT;
+                qint32 y = qint32(j * TILE_HEIGHT);
 
                 for (uint i = 0; i < layer.ncols; i++) {
-                    uint x = i * TILE_WIDTH;
+                    qint32 x = qint32(i * TILE_WIDTH);
 
                     QImage &tile = layer.image_tiles[j][i];
                     if (x + layer.x_offset < MAX_IMAGE_WIDTH &&
@@ -3208,10 +3221,10 @@ void XCFImageFormat::mergeLayerIntoImage(XCFImage &xcf_image)
 #endif
 
     for (uint j = 0; j < layer.nrows; j++) {
-        uint y = j * TILE_HEIGHT;
+        qint32 y = qint32(j * TILE_HEIGHT);
 
         for (uint i = 0; i < layer.ncols; i++) {
-            uint x = i * TILE_WIDTH;
+            qint32 x = qint32(i * TILE_WIDTH);
 
             // This seems the best place to apply the dissolve because it
             // depends on the global position of each tile's
@@ -3851,6 +3864,9 @@ bool XCFImageFormat::mergeGrayAToRGB(const Layer &layer, uint i, uint j, int k, 
     }
 
     switch (layer.mode) {
+    case GIMP_LAYER_MODE_NORMAL:
+    case GIMP_LAYER_MODE_NORMAL_LEGACY:
+        break;
     case GIMP_LAYER_MODE_MULTIPLY:
     case GIMP_LAYER_MODE_MULTIPLY_LEGACY: {
         src = INT_MULT(src, dst);
@@ -4144,7 +4160,9 @@ bool XCFHandler::canRead() const
 bool XCFHandler::read(QImage *image)
 {
     XCFImageFormat xcfif;
-    return xcfif.readXCF(device(), image);
+    auto ok = xcfif.readXCF(device(), image);
+    m_imageSize = image->size();
+    return ok;
 }
 
 bool XCFHandler::write(const QImage &)
@@ -4164,6 +4182,9 @@ QVariant XCFHandler::option(ImageOption option) const
     QVariant v;
 
     if (option == QImageIOHandler::Size) {
+        if (!m_imageSize.isEmpty()) {
+            return m_imageSize;
+        }
         /*
          * The image structure always starts at offset 0 in the XCF file.
          * byte[9]     "gimp xcf " File type identification
@@ -4176,7 +4197,7 @@ QVariant XCFHandler::option(ImageOption option) const
          * uint32      width        Width of canvas
          * uint32      height       Height of canvas
          */
-        if (auto d = device()) {
+        else if (auto d = device()) {
             // transactions works on both random and sequential devices
             d->startTransaction();
             auto ba9 = d->read(9);      // "gimp xcf "
@@ -4209,66 +4230,43 @@ bool XCFHandler::canRead(QIODevice *device)
     }
 
     const qint64 oldPos = device->pos();
-    if (!device->isSequential()) {
-        QDataStream ds(device);
-        XCFImageFormat::XCFImage::Header header;
-        bool failed = !XCFImageFormat::readXCFHeader(ds, &header);
-        ds.setDevice(nullptr);
-        device->seek(oldPos);
-        if (failed) {
-            return false;
-        }
 
-        switch (header.precision) {
-        case XCFImageFormat::GIMP_PRECISION_HALF_LINEAR:
-        case XCFImageFormat::GIMP_PRECISION_HALF_NON_LINEAR:
-        case XCFImageFormat::GIMP_PRECISION_HALF_PERCEPTUAL:
-        case XCFImageFormat::GIMP_PRECISION_FLOAT_LINEAR:
-        case XCFImageFormat::GIMP_PRECISION_FLOAT_NON_LINEAR:
-        case XCFImageFormat::GIMP_PRECISION_FLOAT_PERCEPTUAL:
-        case XCFImageFormat::GIMP_PRECISION_U8_LINEAR:
-        case XCFImageFormat::GIMP_PRECISION_U8_NON_LINEAR:
-        case XCFImageFormat::GIMP_PRECISION_U8_PERCEPTUAL:
-        case XCFImageFormat::GIMP_PRECISION_U16_LINEAR:
-        case XCFImageFormat::GIMP_PRECISION_U16_NON_LINEAR:
-        case XCFImageFormat::GIMP_PRECISION_U16_PERCEPTUAL:
-        case XCFImageFormat::GIMP_PRECISION_U32_LINEAR:
-        case XCFImageFormat::GIMP_PRECISION_U32_NON_LINEAR:
-        case XCFImageFormat::GIMP_PRECISION_U32_PERCEPTUAL:
-            break;
-        case XCFImageFormat::GIMP_PRECISION_DOUBLE_LINEAR:
-        case XCFImageFormat::GIMP_PRECISION_DOUBLE_NON_LINEAR:
-        case XCFImageFormat::GIMP_PRECISION_DOUBLE_PERCEPTUAL:
-        default:
-            qCDebug(XCFPLUGIN) << "unsupported precision" << header.precision;
-            return false;
-        }
+    QDataStream ds(device);
+    XCFImageFormat::XCFImage::Header header;
+    bool failed = !XCFImageFormat::readXCFHeader(ds, &header);
+    ds.setDevice(nullptr);
 
-        return true;
-    }
-
-    char head[8];
-    qint64 readBytes = device->read(head, sizeof(head));
-    if (readBytes != sizeof(head)) {
-        if (device->isSequential()) {
-            while (readBytes > 0) {
-                device->ungetChar(head[readBytes-- - 1]);
-            }
-        } else {
-            device->seek(oldPos);
-        }
+    device->seek(oldPos);
+    if (failed) {
         return false;
     }
 
-    if (device->isSequential()) {
-        while (readBytes > 0) {
-            device->ungetChar(head[readBytes-- - 1]);
-        }
-    } else {
-        device->seek(oldPos);
+    switch (header.precision) {
+    case XCFImageFormat::GIMP_PRECISION_HALF_LINEAR:
+    case XCFImageFormat::GIMP_PRECISION_HALF_NON_LINEAR:
+    case XCFImageFormat::GIMP_PRECISION_HALF_PERCEPTUAL:
+    case XCFImageFormat::GIMP_PRECISION_FLOAT_LINEAR:
+    case XCFImageFormat::GIMP_PRECISION_FLOAT_NON_LINEAR:
+    case XCFImageFormat::GIMP_PRECISION_FLOAT_PERCEPTUAL:
+    case XCFImageFormat::GIMP_PRECISION_U8_LINEAR:
+    case XCFImageFormat::GIMP_PRECISION_U8_NON_LINEAR:
+    case XCFImageFormat::GIMP_PRECISION_U8_PERCEPTUAL:
+    case XCFImageFormat::GIMP_PRECISION_U16_LINEAR:
+    case XCFImageFormat::GIMP_PRECISION_U16_NON_LINEAR:
+    case XCFImageFormat::GIMP_PRECISION_U16_PERCEPTUAL:
+    case XCFImageFormat::GIMP_PRECISION_U32_LINEAR:
+    case XCFImageFormat::GIMP_PRECISION_U32_NON_LINEAR:
+    case XCFImageFormat::GIMP_PRECISION_U32_PERCEPTUAL:
+        break;
+    case XCFImageFormat::GIMP_PRECISION_DOUBLE_LINEAR:
+    case XCFImageFormat::GIMP_PRECISION_DOUBLE_NON_LINEAR:
+    case XCFImageFormat::GIMP_PRECISION_DOUBLE_PERCEPTUAL:
+    default:
+        qCDebug(XCFPLUGIN) << "unsupported precision" << header.precision;
+        return false;
     }
 
-    return qstrncmp(head, "gimp xcf", 8) == 0;
+    return true;
 }
 
 QImageIOPlugin::Capabilities XCFPlugin::capabilities(QIODevice *device, const QByteArray &format) const